+++ lib/roulette.dart
@@ -17,7 +17,12 @@
 library roulette;
 
 export './src/roulette.dart' show Roulette;
-export './src/roulette_controller.dart' show RouletteController;
+export './src/roulette_controller.dart'
+    show
+        RouletteController,
+        AnimationConfig,
+        CurveAnimationConfig,
+        PhysicsAnimationConfig;
 export './src/roulette_group.dart' show RouletteGroup;
 export './src/roulette_unit.dart' show RouletteUnit;
 export './src/roulette_style.dart' show RouletteStyle, SectionImageLayout;